biab/requirements.txt
firebase-admin
google-auth
requests
google-api-python-client